+++ NEWS.md
@@ -1,3 +1,5 @@
++ Changed output of `invExWASDS` function to return a list of tables instead of an `ExposomeSet` object.
+
 # dsExposome 2.0.1