+++ NEWS.md
@@ -1,3 +1,5 @@
+# dsExposome (development version)
+
 # dsExposome 2.0.2
 
 + Changed output of `invExWASDS` function to return a list of tables instead of an `ExposomeSet` object.